refactor(spiffe): extract private-address check into PrivateAddressGuard helper
